Animations (#20)
* added AnimationManager, refactored existing animations

* reorganize files in the folders, now more cleaner includes paths,
refactor move method to pawn,
clean up remaining bonus status values,
clean up remaining anim values in pawn,
clean up unit test, remove todo

* fix cmake build

* refactoring rendering flow,
extract random to utils, sort an initialization list

* refactoring sdl environment, cleanup, fix unit tests

* cmake build fix, vsync on by default

* refactor bot move, now they check all 4 sides and randomly choose between free paths to turnaround

* Add network event replication tests and refactor network handlers

* add two test statistics and fortress change

* add two test BonusSpawn and BonusDeSpawn, fix bonusDeSpawn command serialization

* fix fortress replication

* add two network testm
refactor replace enums with scoped enums

* add two network test,
refactor replace enums with scoped enums

* Refactor animationManager, added animationReplication

* add a new network unit test for respawn tank replication,
fix unit test and fix cmake build.

* fix respawn resource change replication, fix unsubscribing, small refactoring

* fix explosion displaying and scale size, reorganize component into managers folder

* fix unit test after reorganizing the managers folder

* add spawn animation, fix bullet and tank explosion and his positioning (re-center),

* add scaling and centering for explosions, reduce bullet damage radius

* refactoring tank move beh, fix network test, cleanup

* refactoring an event system to deducing types from lambdas, adjust code with a new style without <types>

* fix event system, fix draw object subscription,
(4 network tests still failed, 95 others success)

* fix event system for const char* auto conversion to string

* quick fix event system with std::string_view, improve move using in commands, add no excepts

* more std algorithm and move

* refactored texture and animation manager, add reset for animation manager, fix menu display just in time

* huge refactoring texture and animation manager,
change animation spawn and dispose logic
add rule of five for baseObj and animationObject,
fix animation bugs and crashes

* fix cmake build, cleanup, remove expired TODOs

* fix tank explosion, refactoring creation animation manager, renamings

* refactoring animation flow with interface cleanup, reduce dependency, done few TODO

* refactoring animation flow and refactoring replicate change health, cleanup dependency

* fix build

* add respawn delay, refactoring, extract timer to utils

* fix -1 in respawn point, cleanup bonus interfaces, add new respawn tests

* add texture atlas transparency support

* cleanup

* cleanup bonus interface

* small refactoring animation and cleanup

* Added few if's to handle memory move UB. Rename some functions to make them more readable.

Need to find out where we have memory corruption for Tank_Explosion animations.

* bug fix for rule of 5

* Initialization and assignment of gamepads for player1.

* Revert default value of "isVsyncOn" to "false"

* update dependency and fix cmake build.

* Refactored GetScale and GetCentered into  GetScaledBy. Fixed TankExplosion/BulletExplosion scaling\centering.

* fix multiplayer tank move (client subscribe now not skipped)

* fix multiplayer animation replication for bullets and reformat

* refactor: extract RespawnResourceManager from TankSpawner, adopt designated initializers across,  and code cleanup

* Description: Controllers definition for players and controllers swap.

* New realization of controller's swap with minor ajustments for controllers initilization

* New realization of controller's swap with minor ajustments for controllers initilization - V2

* Restore commented code passage in GameSucess.cpp

* WhiteSpaces_and_Tabs_removement

* reformat

* Restore previously removed header-file

* Refactor of input bindings, plug&play functionality for controllers.

* refatoring key state before sending handling events, restore deleted comments and whitespaces

* reformat

* Fix for gamepads bindings. Definition of player input tags. Gamepads plug&play ajustments.

* fixup! Refactor of input bindings, plug&play functionality for controllers.

* refactoring part of gamesuccess

* Refactoring of keyboard/gamepads control schemes.

* Restore functionality of "Enter"-key in Menu.

* fix build and fix unittests

* refactoring controller input

* omtited std::move for small object

* refactoring behaviors, remove unnesessary std::move, reformat

* fix tank moving ability while active spawn animation

* fix network input

* add flow to client tank enabling

* fix obstacle desync on client at match start

* fix fortress replication and fortress refactoring

* fix bonus helmet status replication

* fix fortrerss unit test and add new test for helmet status replication

* refactoring key state command and cleanup

* Controller adding (#23)

* UserInput event-handlers refactoring (stage one).

* Refactoring of Controllers input functionality

* Fix "unnesessary empty line" and white_space in line #200 in UserInput.cpp

* refactoring SDL environment initialization and controller management

* reformat
refactor controller tag definer
and small refactor client on position change

* aligment fix

* cleanup

* refactoring shared to unique ptr

* refactoring animation objects creating,
fix bug with unnesessary unpausing when window drag in pause, cleanup

* ControllerAdding (#24)

* UserInput event-handlers refactoring (stage one).

* Refactoring of Controllers input functionality

* Fix "unnesessary empty line" and white_space in line #200 in UserInput.cpp

* Add "Menu" and "Pause" buttons functionality for controllers.

* Fix for issue with moving player while menu is active and refactoring of userinput subscriptions.

* Fix for issue with moving enemies and paused players while Menu is opened.

* Revert "Fix for issue with moving player while menu is active and refactoring of userinput subscriptions."

This reverts commit 90250c2c38b98dc7415ca58ef246491be1051166.

* restore after revert

* WIP: state manager and pause dispaying

* fix and some revert after sub merge

* fix pause switch on menu open, reformating and cleanup

* state manager adding gameover and pause refactoring, file renaming

* Made new way to handle "GameOver" state into the SateManager

* add preUpdate and postUpdate stage

* add PreDraw and PostDraw and PreDrawUserInterface and PostDrawUserInterface stages and its fix bugs with overlaping, cleanup game success

* add FramePerSecondManager and refactor GameSuccess

* fix cmake build and unit tests

* refatoring animation manager, cleanup dependencies

* add spawnManager and fix unit tests, refactoring obstacle spawner and map loading

* resolve review issues

* add renderManager, refactor rendering logic so sld rendered in one class, cleanup unused interfaces

* fix color in menu white->cyan and refactoring int color to unsigned int, reformat

* Add WinCondition and visual placeholder for it.

* WinCondition refactoring to be alligned with RenderManager.

* fix recenter health bar when pickup star bonus and fix some warnings

* fix for fps dispaying and fps drop, cleanup

* refactoring fps count, fix cmake build, deltatime now double, cleanup

* Refactor of wincondition functionality.

* renamings, reformat

* add bonus(start, caliber, tank) network handling, cleanup outdated server stuff and unify parameter usage

* refactor statistics, fix unsubscribe, cleanup

* fix unsubscribe and reorder, cleanup

* reformat

* fix build, after removing SDL_Log dependency

* Remake "GameOver" logic in "GameWon"-like style.

* Deletion of unnecessary GameOver-related conditions in StateManager.

* Refactoring of Win/Loose functionality.

* One more refactoring for Win/Loose functionality.

* refactor win\lose condition

---------


---------

+++ headers/interfaces/ITickUpdatable.h
@@ -2,7 +2,7 @@
 
 class ITickUpdatable
 {
-	virtual void TickUpdate(float deltaTime) = 0;
+	virtual void TickUpdate(double deltaTime) = 0;
 
 protected:
 	virtual ~ITickUpdatable() = default;
